feature: refactor sign-in process and add leaderboard functionality
