+++ app/src/main/java/com/dudoji/tangvivor/matching/activity/RoomListActivity.kt
@@ -1,5 +1,6 @@
 package com.dudoji.tangvivor.matching.activity
 
+import RoomListAdapter
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -13,7 +14,6 @@ import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.RecyclerView
 import com.dudoji.tangvivor.R
 import com.dudoji.tangvivor.game.GameActivity
-import com.dudoji.tangvivor.matching.entity.Room
 import com.dudoji.tangvivor.repository.GameRepository
 import com.dudoji.tangvivor.repository.RoomRepository
 import com.dudoji.tangvivor.repository.UserRepository
@@ -29,33 +29,18 @@ class RoomListActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_room_list)
         findViewById<ImageButton>(R.id.leader_board_button).setOnClickListener{
-            PlayGames.getLeaderboardsClient(this)
-                .getLeaderboardIntent(getString(R.string.score_leaderboard_id))
-                .addOnSuccessListener(this){intent ->
-                    startActivityForResult(intent, 1001)
-                }
-                .addOnFailureListener(this) { e ->
-                    Log.e("RoomListActivity", "Failed to get leaderboard intent", e)
-                    Toast.makeText(this, "Failed to open leaderboard", Toast.LENGTH_SHORT).show()
-                }
+            openLeaderboard()
         }
 
         findViewById<TextView>(R.id.my_name).text = UserRepository.me?.name
 
         roomListRecyclerView = findViewById(R.id.room_list_recycler_view)
         roomListRecyclerView.layoutManager = androidx.recyclerview.widget.LinearLayoutManager(this)
-        lifecycleScope.launch {
-            val roomList = RoomRepository.getRooms()
-            Log.d("RoomListActivity", "Room list fetched: $roomList")
-            roomListRecyclerView.adapter = RoomListAdapter(roomList, this@RoomListActivity)
-        }
+        checkMatching()
 
         reloadButton = findViewById(R.id.reload_button)
         reloadButton.setOnClickListener {
-            lifecycleScope.launch {
-                val roomList = RoomRepository.getRooms()
-                roomListRecyclerView.adapter = RoomListAdapter(roomList, this@RoomListActivity)
-            }
+            checkMatching()
         }
         roomNameEdit = findViewById(R.id.room_name)
         createRoomButton = findViewById(R.id.create_room_button)
@@ -74,65 +59,23 @@ class RoomListActivity : ComponentActivity() {
         }
     }
 
+    fun openLeaderboard() {
+        PlayGames.getLeaderboardsClient(this)
+            .getLeaderboardIntent(getString(R.string.score_leaderboard_id))
+            .addOnSuccessListener(this){intent ->
+                startActivityForResult(intent, 1001)
+            }
+            .addOnFailureListener(this) { e ->
+                Log.e("RoomListActivity", "Failed to get leaderboard intent", e)
+                Toast.makeText(this, "Failed to open leaderboard", Toast.LENGTH_SHORT).show()
+            }
+    }
+
     fun checkMatching() {
         lifecycleScope.launch {
             val roomList = RoomRepository.getRooms()
             Log.d("RoomListActivity", "Checking matching: $roomList")
             roomListRecyclerView.adapter = RoomListAdapter(roomList, this@RoomListActivity)
         }
-    }
-}
-
-class RoomListAdapter(val roomList: List<Room>, val activity: RoomListActivity) : RecyclerView.Adapter<RoomListAdapter.RoomViewHolder>() {
-
-    class RoomViewHolder(itemView: android.view.View) : RecyclerView.ViewHolder(itemView) {
-        val nameTextView = itemView.findViewById<android.widget.TextView>(R.id.item_name)
-        val makerTextView = itemView.findViewById<android.widget.TextView>(R.id.item_maker)
-    }
-
-    override fun onCreateViewHolder(parent: android.view.ViewGroup, viewType: Int): RoomViewHolder {
-        val itemView = android.view.LayoutInflater.from(parent.context)
-            .inflate(R.layout.game_item, parent, false)
-        return RoomViewHolder(itemView)
-    }
-
-    override fun onBindViewHolder(holder: RoomViewHolder, position: Int) {
-        val room = roomList[position]
-        holder.nameTextView.text = room.name
-        activity.lifecycleScope.launch {
-            holder.makerTextView.text = UserRepository.getUser(room.user1!!).name
-        }
-
-        holder.itemView.setOnClickListener {
-            RoomRepository.db.collection(RoomRepository.COLLECTION_NAME)
-                .document(room.name!!)
-                .update("user2", UserRepository.me?.id)
-                .addOnSuccessListener {
-                    Toast.makeText(
-                        holder.itemView.context,
-                        "Joined room: ${room.name}",
-                        Toast.LENGTH_SHORT
-                    ).show()
-                    RoomRepository.db.collection(RoomRepository.COLLECTION_NAME)
-                        .document(room.name!!)
-                        .delete()
-                    val intent = Intent(holder.itemView.context, com.dudoji.tangvivor.game.GameActivity::class.java)
-                    intent.putExtra("roomName", room.name)
-                    intent.putExtra("me", 2)
-                    holder.itemView.context.startActivity(intent)
-                }
-                .addOnFailureListener { e ->
-                    Log.d("RoomListAdapter", "Error joining room: ", e)
-                    Toast.makeText(
-                        holder.itemView.context,
-                        "Failed to join room",
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-        }
-    }
-
-    override fun getItemCount(): Int {
-        return roomList.size
     }
 }
